Refator landscape view manager.Add hooks to folder
