+++ src/components/landscape-view-manager/landscape-view-manager.js
@@ -1,10 +1,3 @@
-import { connect } from 'react-redux';
 import Component from './landscape-view-manager-component';
-// This import forces the registering of the module
-import 'redux_modules/grid-cell-data';
-import 'redux_modules/geo-description';
 
-import * as actions from 'redux_modules/grid-cell-data/grid-cell-data-actions';
-import { fetchGeoDescription } from 'redux_modules/geo-description/geo-description-actions';
-
-export default connect(null, { ...actions, fetchGeoDescription })(Component);
+export default Component;
