+++ src/constants/landscape-view-constants.js
@@ -4,3 +4,5 @@ export const gridCellDefaultStyles = {
   outlineWidth: 2,
   colorRGB: [147, 255, 95]
 }
+
+export const ZOOM_LEVEL_TRIGGER = 8;
